+++ mock_data_gen_cloud_run_function/requirements.txt
@@ -1,0 +1,4 @@
+Faker
+google-cloud-storage
+google-cloud-pubsub
+Flask
