+++ pyspark_src/main.py
@@ -2,10 +2,10 @@ import random
 from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.functions import  *
 
-from spark_builder import *
-from read_data import *
-from pydeequ_checks import *
-from writer import *
+from modules.spark_builder import *
+from modules.read_data import *
+from modules.pydeequ_checks import *
+from modules.writer import *
 
 from google.cloud import bigquery
 from google.cloud import storage
@@ -14,14 +14,14 @@ from google.cloud import storage
 PROJECT_ID = "dataproc-spark-461405"
 BUCKET_NAME = 'transactions_raw_data'
 ARCHIVE_BUCKET = 'transactions_raw_data_archive'
-GCS_client = storage.Client()
+GCS_client = storage.Client(project=PROJECT_ID)
 
 # BigQuery variables
 DATA_SET = 'financial_transactions'
 TRANSACTIONS_TABLE =  'daily_transactions'
 CUSTOMER_INFO_STAGING_TABLE = 'stg_daily_customer_info'
 CUSTOMER_INFO_TARGET_TABLE = 'customer_profile_scd2'
-bq_client = bigquery.Client()
+bq_client = bigquery.Client(project=PROJECT_ID)
 
 def main():
     # 1 --> Create a Spark Session and fetch dates from raw data bucket---
@@ -29,9 +29,9 @@ def main():
     dates = list_gcs_files_dates(BUCKET_NAME, GCS_client)
 
     if dates:
-        for date in dates:
+        for date in dates:    
 
-            local_test_path =  f"/Users/delgadonoriega/Desktop/gcp-data-eng-bootcamp/Module_3_class_1/pyspark_src/transactions_{date}.json"  # Pending code to build GCS path
+            local_test_path = f"/Users/delgadonoriega/Desktop/gcp-data-eng-bootcamp/Module_3_class_1/financial_transactions_project/pyspark_src/transactions_{date}.json"
             # gcs_path =f"gs://{BUCKET_NAME}/transactions_{date}.json"
             # gcs_path_archive = f"gs://{ARCHIVE_BUCKET}/transactions_{date}.json"
 
@@ -51,11 +51,11 @@ def main():
             run_quality_checks(spark, data)
 
             # 4 --> Write all daily to transactions table
-            # write_to_transactions_table(PROJECT_ID, DATA_SET, TRANSACTIONS_TABLE, transactions)
+            # pyspark_to_bigquery(PROJECT_ID, DATA_SET, TRANSACTIONS_TABLE, transactions, 'append')
             big_query_write_test(project_id=PROJECT_ID, dataset=DATA_SET, table=TRANSACTIONS_TABLE, BQ_client=bq_client, df=transactions, overwrite=False)
             
             # 5 --> Write to stagin table table
-            # write_to_transactions_table(PROJECT_ID, DATA_SET, CUSTOMER_INFO_STAGING_TABLE, customer_info)
+            # pyspark_to_bigquery(PROJECT_ID, DATA_SET, CUSTOMER_INFO_STAGING_TABLE, customer_info, 'overwrite')
             big_query_write_test(project_id=PROJECT_ID, dataset=DATA_SET, table=CUSTOMER_INFO_STAGING_TABLE, BQ_client=bq_client, df=customer_info, overwrite=True)
 
             # 6 --> SCD2 with target table
